Added the ability to change player skin
Added the ability to change the player skin from the sidebar settings panel. Now users will be able to select different skins from the predefined list of skins (4 at the moment) and if they want to use any custom skin they can do so by toggling the swicth on and adding any custom skin URL in the input fields. Also added the thumbnails for the predefined skins for better user experience.